Update: Qna/follow/seller entity update

+++ src/main/java/com/myomi/qna/repository/QnaRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.myomi.qna.entity.Qna;
 
-public interface QnaRepository extends CrudRepository<Qna, Integer> {
+
+public interface QnaRepository extends CrudRepository<Qna, Long> {
 
 }
